新增jd商品api

+++ apis/jd-api/src/main/java/cn/qihangerp/api/jd/bo/PullRequest.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class PullRequest {
-    private Integer shopId;//店铺Id
+    private Long shopId;//店铺Id
     private String orderId;
 }
